# Improve the performance of `softmax_forward8` in `dev/cuda`