@KarhouTam 2024-06-10
I agree that `__shfl_xor_sync()` is way complicated than  `__shfl_down_sync()`.

In my view, if just for reduction, `__shfl_down_sync()` and `__shfl_xor_sync()` can properly do the job, and I found that using `__shfl_donw_sync()` for reduction needs `val = __shfl_sync(0xFFFFFFFF, val, 0)` for synchronization that `__shfl_xor_sync()` doesn't. I'm still finding the specific reason behind this. 

Here is the comparison. These changes don't improve the kernel that much, but they are not harmful, as least I think that.

### Comparison Before and After Changes
#### Using `__shfl_down_sync()` with `__shfl_sync()`
```
block_size   32 | time 15.2935 ms | per token 1.87 µs
block_size   64 | time 15.2867 ms | per token 1.87 µs
block_size  128 | time 17.3360 ms | per token 2.12 µs
block_size  256 | time 19.3772 ms | per token 2.37 µs
block_size  512 | time 21.4683 ms | per token 2.62 µs
block_size 1024 | time 25.6833 ms | per token 3.14 µs
```

#### Using `__shfl_xor_sync()` without `__shfl_sync()`
```
block_size   32 | time 14.1287 ms | per token 1.72 µs
block_size   64 | time 13.2909 ms | per token 1.62 µs
block_size  128 | time 13.8408 ms | per token 1.69 µs
block_size  256 | time 16.6469 ms | per token 2.03 µs
block_size  512 | time 18.0218 ms | per token 2.20 µs
block_size 1024 | time 19.5242 ms | per token 2.38 µs
```

Thanks for your code review.
I'm just a cuda newbie and trying to participate some open source project to improve myself. 
Respect you and your contribution 🙌.