@ngc92 2024-06-10
xor shuffle is the way to go if you need the result in all threads. What it implements is a butterfly-shuffle type operation, that ensures that in the end, every thread actually computes the reduction over all values. Its actually not that complicated once you visualize what is going on, e.g., [this picture](https://www.researchgate.net/figure/Data-exchange-among-8-threads-using-butterfly-warp-shuffle-operations_fig1_317485271) was the first that came up on google. Then, of course, you need to convince yourself that the xor operation actually does create this pattern.
(In contrast, shfl down would have all  the arrows pointing to the left, so that only the leftmost element would actually get the full reduction; that's why you need one more shuffle to broadcast the results)

Because it is so useful but at the same time non-trivial, we actually do have a helper function in `common.h` that implements exactly this operation: `warpReduceSum`.

The changes actually give more improvement than I would have guessed  :)