@KarhouTam 2024-06-10
Thank you for your comprehensive explanation. It really helps me understand these warp reduce functions.

I do know llm.c provides warp reduce API in common.h. Considering the reduction in online softmax requires additional comparisons, I haven't used them, which are suitable for more general cases.

BTW, I see that in most warp reduce API implementations, the offset is mostly chagned by `offset /=2`. Consider the warp size in most scenarios is fixed 32, will it more efficient that using `offset >= 1` instead? Because in common sense, bitwise operation is more computationally lightweight than division. Or compiler actually do optimization for us and there is no difference? 

Thank you for your reviewing and comment again.